refatorando os teste
